@@ -24,14 +24,12 @@
 //
 #if BLOCK_DEFINITION_MODE == BLOCK_ENUM_MODE
 
-//#define START_BLOCK_DEFINITION(__NAME__, __SIZE__, __COLOR__) __NAME__##_BLOCK_TYPE,
 #define START_BLOCK_DEFINITION(__NAME__, __SIZE__, __COLOR__) static const Type __NAME__;
 #define ADD_FACE_CODE(__WHICHFACE__, __SIZE__, __CODE__, __DOCK_ORIENTATIONS__)
 #define END_BLOCK_DEFINITION
 
 #elif BLOCK_DEFINITION_MODE == BLOCK_ENUM_VALUE_MODE
 
-//#define START_BLOCK_DEFINITION(__NAME__, __SIZE__, __COLOR__) __NAME__##_BLOCK_TYPE,
 #define START_BLOCK_DEFINITION(__NAME__, __SIZE__, __COLOR__) const Block::Type Block::Type::__NAME__(QUOTE(__NAME__));
 #define ADD_FACE_CODE(__WHICHFACE__, __SIZE__, __CODE__, __DOCK_ORIENTATIONS__)
 #define END_BLOCK_DEFINITION
@@ -46,7 +44,11 @@
 {Block::Type::Block_##__NAME__, {.name = QUOTE(__NAME__), .color = ::Anki::NamedColors::__COLOR__, .size = {UNWRAP __SIZE__}, .faces = {
 
 #define ADD_FACE_CODE(__WHICHFACE__, __SIZE__, __CODE__, __DOCK_ORIENTATIONS__) \
-{.whichFace = __WHICHFACE__, .size = __SIZE__, .code = __CODE__, .dockOrientations = __DOCK_ORIENTATIONS__, .rollOrientations = PreActionOrientation::ALL},
+{.whichFace        = __WHICHFACE__, \
+ .code             = __CODE__,      \
+ .size             = __SIZE__,      \
+ .dockOrientations = __DOCK_ORIENTATIONS__, \
+ .rollOrientations = PreActionOrientation::ALL},
 
 #define END_BLOCK_DEFINITION  } } },
 
@@ -59,13 +61,9 @@
 #define ADD_FACE_CODE(__WHICHFACE__, __SIZE__, __CODE__, __DOCK_ORIENTATIONS__)
 #define END_BLOCK_DEFINITION
 
-//
-// Unknown mode! (Error)
-//
 #else
 #error Unknown BLOCK_DEFINITION_MODE!
-
-#endif // #if/elif BLOCK_DEFINITION_MODE == <which_mode>
+#endif
 
 #define ADD_ALL_FACES(__SIZE__, __CODE__) \
 ADD_FACE_CODE(FRONT_FACE,  __SIZE__, __CODE__, PreActionOrientation::ALL) \
@@ -77,4 +75,4 @@ ADD_FACE_CODE(BOTTOM_FACE, __SIZE__, __CODE__, PreActionOrientation::ALL)
 
 #undef BLOCK_DEFINITION_MODE
 
-#endif // #ifndef BLOCK_DEFINTION_MODE
+#endif // BLOCK_DEFINITION_MODE
