@@ -79,8 +79,8 @@ void ConditionObjectPositionUpdated::HandleEvent(const EngineToGameEvent& event,
     
     default:
     PRINT_NAMED_ERROR("BehaviorAcknowledgeObject.HandleWhileNotRunning.InvalidTag",
-                      "Received event with unhandled tag %hhu.",
-                      event.GetData().GetTag());
+                      "Received event with unhandled tag %u.",
+                      static_cast<uint32_t>(event.GetData().GetTag()));
     break;
   }
 }
