@@ -23,7 +23,7 @@
 //extern "C" {
 //#endif
 
-//size_t strlcpy(char *, const char *, size_t);
+//extern size_t strlcpy(char *, const char *, size_t);
 
 //#ifdef __cplusplus
 //}
