@@ -129,7 +129,8 @@ std::string StringMapToPrettyJson(const std::map<std::string,std::string> &strin
 void StringTrimWhitespaceFromEnd(std::string& s)
 {
   s.erase(std::find_if(s.rbegin(), s.rend(),
-                       std::not1(std::ptr_fun<int, int>(std::isspace))).base(), s.end());
+                       [](int c){ return !std::isspace(c); }).base(),
+          s.end());
 }
 
 } // namespace AnkiUtil
