@@ -210,15 +210,17 @@ bool StringEndsWith(const std::string& fullString, const std::string& ending)
 void StringTrimWhitespaceFromEnd(std::string& s)
 {
   s.erase(std::find_if(s.rbegin(), s.rend(),
-                       std::not1(std::ptr_fun<int, int>(std::isspace))).base(), s.end());
+                       [](int c){ return !std::isspace(c); }).base(),
+          s.end());
 }
 
-void StringTrimWhitespaceFromStart(std::string& s) {
-    s.erase(s.begin(),
-            std::find_if(s.begin(),
-                         s.end(),
-                         std::not1(std::ptr_fun<int, int>(std::isspace))));
+void StringTrimWhitespaceFromStart(std::string& s)
+{
+  s.erase(s.begin(),
+          std::find_if(s.begin(), s.end(),
+                       [](int c){ return !std::isspace(c); }));
 }
+
 
 void StringTrimWhitespace(std::string& s) {
   StringTrimWhitespaceFromStart(s);
